Implement IP-packet handling code.
This commit contains Bowen's work on the ip-packet handling branch.
Unfortunately, his 'ip-branch' was not able to be cleanly applied;
Rather than being merged into the branch, code was copy-pasted in from
various other branches in different stages of development. Because of
this, the branch's code did not reflect the state of development. As
such, I have taken the novel aspects of Bowen's IP-packet handling code
and moved them unaltered to this branch for development and testing.

This 'merges' in most of Bowen's code from his 'ip-branch' as of commit
da17cad4a50545214794b7b26f41aff805f5424e.